MATE-10 : [REFACTOR] Jwt 토큰 응답 DTO 클래스 리팩토링
- 새로 가입한 회원 여부를 반환하는 isNewMember 필드 추가
