+++ src/main/java/com/example/mate/domain/auth/controller/AuthController.java
@@ -1,5 +1,6 @@
 package com.example.mate.domain.auth.controller;
 
+import com.example.mate.domain.member.dto.response.LoginTokenResponse;
 import com.example.mate.global.jwt.JwtToken;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseEntity;
@@ -22,14 +23,15 @@ public class AuthController {
     4. 사용자 정보 Jwt 토큰으로 반환
     */
     @GetMapping("/login/naver")
-    public ResponseEntity<JwtToken> loginByNaver(@RequestParam String code) {
-        JwtToken jwtToken = JwtToken.builder()
+    public ResponseEntity<LoginTokenResponse> loginByNaver(@RequestParam String code) {
+        LoginTokenResponse response = LoginTokenResponse.builder()
                 .grantType("Bearer")
                 .accessToken("accessToken")
                 .refreshToken("refreshToken")
+                .isNewMember(true)
                 .build();
 
-        return ResponseEntity.ok(jwtToken);
+        return ResponseEntity.ok(response);
     }
 
     /*
@@ -40,14 +42,15 @@ public class AuthController {
     4. 사용자 정보 Jwt 토큰으로 반환
     */
     @GetMapping("/login/google")
-    public ResponseEntity<JwtToken> loginByGoogle(@RequestParam String code) {
-        JwtToken jwtToken = JwtToken.builder()
+    public ResponseEntity<LoginTokenResponse> loginByGoogle(@RequestParam String code) {
+        LoginTokenResponse response = LoginTokenResponse.builder()
                 .grantType("Bearer")
                 .accessToken("accessToken")
                 .refreshToken("refreshToken")
+                .isNewMember(true)
                 .build();
 
-        return ResponseEntity.ok(jwtToken);
+        return ResponseEntity.ok(response);
     }
 
     /*
